backend/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
sqlalchemy
pydantic
python-dotenv
openai
httpx
python-multipart
